Portfolio Website
I created a portfolio website with a more detailed version of my resume. On this site, there is a brief information about myself, a contact form where you can reach me, a list of my talents and skills, and a link to download my resume. It is still under construction.